rename pkg to cluster package
