+++ cluster/status.go
@@ -6,11 +6,11 @@
 // $ goagen
 // --design=github.com/fabric8-services/fabric8-cluster/design
 // --out=$(GOPATH)/src/github.com/fabric8-services/fabric8-cluster-client
-// --pkg=pkg
+// --pkg=cluster
 // --tool=cluster
 // --version=v1.3.0
 
-package pkg
+package cluster
 
 import (
 	"context"
